My First version...
As a quick side note, I've built this around my use with a Microsoft
SideWinder Steering Wheel and Pedals (USB, V1) and tested with Forza
Horizon 3  -  Working perfectly! :+1:

- Fixed bug/issue where camera in certain games was in a permanently
Panned state usually to the front-left.
-- This was caused by un-assigned axes still getting a value rather than
0.

- Fixed a bug when ControllerOptions.cs was looping around the available
Axes count and not the actual Axes values (e.g. Brake pedal on MS
Sidewinder Force Feedback steering wheel is Axis 6, not Axis 3 but is
The 3rd Axis...)

- Added deadzone/compensation feature, you can add a value onto the raw
Value from the LH-X Axis to compensate in certain games against the
deadzone...
-- I’ve found setting this between 30 and 40 in Forza Horizon 3 makes
for a nice experience… (my current is 40)